Added delete messages

+++ src/ComponentsDeepika/ViewDepartment.jsx
@@ -4,7 +4,23 @@ import { Input } from 'antd';
 import { Button } from 'antd';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
+import moment from "moment";
+import { ExclamationCircleOutlined, PlusOutlined } from "@ant-design/icons";
+import {
+    Modal,
+    notification,
+    Drawer,
+    Form,
+    Col,
+    Row,
+    Select,
+    DatePicker,
+    Popconfirm,
+    message,
+  } from "antd";
 
+  const { confirm } = Modal;
+  const { Option } = Select;
 const { Search } = Input;
 
 class ViewDepartment extends Component {
@@ -31,18 +47,62 @@ class ViewDepartment extends Component {
         
     }
 
-    deletedept = (id) =>{
-        var url 
-        url = "http://127.0.0.1:8000/department-Delete/" + id + "/";
-        fetch(url,{
-            method:"DELETE",
-            headers:{
-                'Content-type':'application/json',
-            },
-        }).then((response) => {
-            this.fetchDepartment()
-        }).catch(err => console.log(err))
-    }
+    //took
+    showDeleteConfirm = (data) => {
+        confirm({
+          title: "Are you sure delete this Department?",
+          icon: <ExclamationCircleOutlined />,
+          content:
+            "You are gonna delete " +
+            data +
+            " from the database. Once you click on the delete button, the deletion can not reverse back ⚠⚠",
+          okText: "Yes,Delete",
+          okType: "danger",
+          cancelText: "No,Cancel",
+          onOk() {
+           
+            const args = {
+              message: "Notification",
+              description: "You have successfully deleted Department " + data,
+              duration: 0,
+            };
+            notification.open(args);
+          },
+          onCancel() {
+            console.log("Cancel");
+          },
+        });
+      };
+    //tookend
+
+
+
+    confirm = (data) => {
+        var url = "http://127.0.0.1:8000/department-Delete/" + data + "/";
+    
+        fetch(url, {
+          method: "DELETE",
+          headers: {
+            "Content-type": "application/json",
+          },
+        })
+          
+          .catch((err) => console.log(err));
+        console.log(data);
+        const args = {
+          message: "Notification",
+          description: "You have successfully deleted department " + data,
+          duration: 0,
+        };
+        notification.open(args);
+      };
+    
+      cancel = (e) => {
+        console.log(e);
+        message.error("Canceled deleting Department");
+      };
+
+
 
     render() { 
         return ( 
@@ -104,20 +164,28 @@ class ViewDepartment extends Component {
                                         >Edit</Button>
                                     </td>
                                     <td>
-                                        <Button type="button"  onClick={() => this.deletedept(user.dept_id)} className="btn btn-danger"
-                                        style={{
-                                            color:"white",
-                                            padding:"5px 10px",
-                                            fontSize:"10px",
-                                            fontWeight:"bold",
-                                            cursor:"pointer",
-                                            backgroundColor:"white",
-                                            color:"black",
-                                            border:"2px solid red",
+                                    <Button
+                                    style={{
+                                        color:"white",
+                                        padding:"5px 10px",
+                                        fontSize:"10px",
+                                        fontWeight:"bold",
+                                        cursor:"pointer",
+                                        backgroundColor:"white",
+                                        color:"black",
+                                        border:"2px solid red",
 
-                                        }}>
-                                        Delete
-                                        </Button>
+                                    }}>
+                                    <Popconfirm
+                                        title="Are you sure delete this Department?"
+                                        onConfirm={() => this.confirm(user.dept_id)}
+                                        onCancel={this.cancel}
+                                        okText="Yes,Delete"
+                                        cancelText="No,Cancel"
+                                    >
+                                        <a href="#">Delete</a>
+                                    </Popconfirm>
+                                    </Button>
                                     </td>
                                     </tr>
                                     );
